requirements.txt
annotated-types==0.7.0
annoy==1.17.3
anyio==4.7.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
executing==2.1.0
fastjsonschema==2.21.1
feather-format==0.4.1
fishersapi==1.0
flask==3.0.3
folium==0.19.4
fonttools==4.53.0
fqdn==1.5.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
hierdiff==0.85
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.34.1
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kiwisolver==1.4.5
lazy-loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.3.2
notebook-shim==0.2.4
numba==0.60.0
numpy==2.0.0
olga==1.2.4
openai==1.58.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
palmotif==0.4
pandas==2.2.2
pandocfilters==1.5.1
parasail==1.3.4
parmap==1.7.0
parso==0.8.4
patsy==1.0.1
pillow==10.3.0
pip==24.3.1
platformdirs==4.3.6
plotly==5.24.1
progress==1.6
prometheus-client==0.21.1
prompt-toolkit==3.0.48
psutil==6.1.0
pure-eval==0.2.3
pwseqdist==0.6
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic-core==2.27.2
pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qgrid==1.3.1
ratelim==0.1.6
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
scrublet==0.2.3
seaborn==0.13.2
selenium==4.27.1
Send2Trash==1.8.3
setuptools==69.5.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
svgwrite==1.4.3
tcrdist3==0.3
tcrsampler==0.1.9
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.6.18
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
undetected-chromedriver==3.5.5
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
werkzeug==3.0.6
widgetsnbextension==4.0.13
wsproto==1.2.0
xyzservices==2024.9.0
zipdist==0.1.5
zipp==3.21.0